RDBC-642 CachingOfDocumentIncludeTest::can_include_nested_paths
